# schema/openapi.yaml
openapi: '3.0.0'
info:
  version: 1.0.0
  title: Wealth API
servers:
  - url: /api
paths:
...
  /watchlist:
    get:
      summary: Get users watchlist
      responses:
        '200':
          description: Valid request
          content:
            application/json:
              schema:
                type: object
    post:
      summary: Add item to watchlist
      requestBody:
        content:
          application/json:
            schema:
              additionalProperties: false
              required:
                - symbol
                - fullName
                - assetType
              properties:
                symbol:
                  type: string
                assetType:
                  type: string
                fullName:
                  type: string
      responses:
        '200':
          description: Valid request
          content:
            application/json:
              schema:
                type: object
    delete:
      summary: Delete watchlist item from watchlist
      parameters:
        - name: itemID
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Valid request
          content:
            application/json:
              schema:
                type: object
...